remove parser module,move to appropriate locations
