+++ midix/Cargo.toml
@@ -12,7 +12,6 @@ categories = ["multimedia::audio", "multimedia::encoding", "multimedia"]
 license = "MIT OR Apache-2.0"
 
 [dependencies]
-thiserror = "2.0.9"
 
 [features]
 nightly = []
