+++ midix/src/file/chunk.rs
@@ -4,7 +4,7 @@
 /// which is the number of bytes in the chunk. This structure allows
 /// future chunk types to be designed which may be easily be ignored
 /// if encountered by a program written before the chunk type is introduced.
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Copy, Debug, Clone, PartialEq, Eq)]
 pub enum MidiChunkType {
     /// Represents the byte length of the midi header.
     ///
